edits w.r.t. review on PR #1
