+++ src/test/java/de/rondiplomatico/nds/NDSPolyFillDemoTest.java
@@ -1,0 +1,13 @@
+package de.rondiplomatico.nds;
+
+import org.junit.Test;
+
+import de.rondiplomatico.nds.NDSPolyFillDemo;
+
+public class NDSPolyFillDemoTest {
+    @Test
+    public void testPolyFillDemoWorks() {
+        String[] test                       = {"Test", "Test2"};
+        NDSPolyFillDemo.main(test);
+    }
+}
