+++ src/test/java/de/rondiplomatico/nds/NDSTileTest.java
@@ -19,6 +19,7 @@ import de.rondiplomatico.nds.WGS84Coordinate;
  * 
  *
  * @author Daniel Wirtz
+ * @author Andreas Hessenthaler
  * @since 20.02.2020
  */
 public class NDSTileTest {
@@ -301,125 +302,33 @@ public class NDSTileTest {
     }
 
     @Test
-    public void testCornersWorks() {
-        // Tile 0
-        NDSTile t               = new NDSTile(0, 0);
-        NDSCoordinate c         = t.getCenter();
-        NDSCoordinate[] corners = t.getCorners();
-        NDSCoordinate sw        = corners[0];
-        NDSCoordinate se        = corners[1];
-        NDSCoordinate ne        = corners[2];
-        NDSCoordinate nw        = corners[3];
-        System.out.println(sw.toWGS84());
-        System.out.println(se.toWGS84());
-        System.out.println(nw.toWGS84());
-        System.out.println(ne.toWGS84());
-        assertEquals(  0.0, sw.toWGS84().getLongitude(), eps);
-        assertEquals(180.0, se.toWGS84().getLongitude(), eps);
-        assertEquals(  0.0, nw.toWGS84().getLongitude(), eps);
-        assertEquals(180.0, ne.toWGS84().getLongitude(), eps);
-        assertEquals(-90.0, sw.toWGS84().getLatitude(),  eps);
-        assertEquals(-90.0, se.toWGS84().getLatitude(),  eps);
-        assertEquals( 90.0, nw.toWGS84().getLatitude(),  eps);
-        assertEquals( 90.0, ne.toWGS84().getLatitude(),  eps);
-        // Tile 1
-        t               = new NDSTile(0, 1);
-        c         = t.getCenter();
-        corners = t.getCorners();
-        sw        = corners[0];
-        se        = corners[1];
-        ne        = corners[2];
-        nw        = corners[3];
-        System.out.println(sw.toWGS84());
-        System.out.println(se.toWGS84());
-        System.out.println(nw.toWGS84());
-        System.out.println(ne.toWGS84());
-        assertEquals(-180.0, sw.toWGS84().getLongitude(), eps);
-        assertEquals(   0.0, se.toWGS84().getLongitude(), eps);
-        assertEquals(-180.0, nw.toWGS84().getLongitude(), eps);
-        assertEquals(   0.0, ne.toWGS84().getLongitude(), eps);
-        assertEquals(-90.0, sw.toWGS84().getLatitude(),  eps);
-        assertEquals(-90.0, se.toWGS84().getLatitude(),  eps);
-        assertEquals( 90.0, nw.toWGS84().getLatitude(),  eps);
-        assertEquals( 90.0, ne.toWGS84().getLatitude(),  eps);
-    }
-
-    @Test
-    public void testMidpointsWorks() {
-        // Tile 0
-        NDSTile t               = new NDSTile(0, 0);
-        NDSCoordinate c         = t.getCenter();
-        NDSCoordinate[] corners = t.getCorners();
-        NDSCoordinate sw        = corners[0];
-        NDSCoordinate se        = corners[1];
-        NDSCoordinate ne        = corners[2];
-        NDSCoordinate nw        = corners[3];
-        NDSCoordinate csw       = sw.getMidpoint(c);
-        NDSCoordinate cse       = se.getMidpoint(c);
-        NDSCoordinate cne       = ne.getMidpoint(c);
-        NDSCoordinate cnw       = nw.getMidpoint(c);
-        assertEquals( 45.0, csw.toWGS84().getLongitude(), eps);
-        assertEquals(135.0, cse.toWGS84().getLongitude(), eps);
-        assertEquals( 45.0, cnw.toWGS84().getLongitude(), eps);
-        assertEquals(135.0, cne.toWGS84().getLongitude(), eps);
-        assertEquals(-45.0, csw.toWGS84().getLatitude(),  eps);
-        assertEquals(-45.0, cse.toWGS84().getLatitude(),  eps);
-        assertEquals( 45.0, cnw.toWGS84().getLatitude(),  eps);
-        assertEquals( 45.0, cne.toWGS84().getLatitude(),  eps);
-        // Tile 1
-        t                       = new NDSTile(0, 1);
-        c                       = t.getCenter();
-        corners                 = t.getCorners();
-        sw                      = corners[0];
-        se                      = corners[1];
-        ne                      = corners[2];
-        nw                      = corners[3];
-        csw                     = sw.getMidpoint(c);
-        cse                     = se.getMidpoint(c);
-        cne                     = ne.getMidpoint(c);
-        cnw                     = nw.getMidpoint(c);
-        assertEquals(-135.0, csw.toWGS84().getLongitude(), eps);
-        assertEquals( -45.0, cse.toWGS84().getLongitude(), eps);
-        assertEquals(-135.0, cnw.toWGS84().getLongitude(), eps);
-        assertEquals( -45.0, cne.toWGS84().getLongitude(), eps);
-        assertEquals( -45.0, csw.toWGS84().getLatitude(),  eps);
-        assertEquals( -45.0, cse.toWGS84().getLatitude(),  eps);
-        assertEquals(  45.0, cnw.toWGS84().getLatitude(),  eps);
-        assertEquals(  45.0, cne.toWGS84().getLatitude(),  eps);
-    }
-
-    @Test
-    public void testChildTileNumbersWorks() {
+    public void testChildTileWorks() {
         // get tile and its center
         int masterTileLevel                     = 3;
         NDSTile t                               = new NDSTile(masterTileLevel, 8);
         NDSCoordinate c                         = t.getCenter();
-        // store the master tile corners; ordering is: sw se ne nw
-        NDSCoordinate[] corners                 = t.getCorners();
-        NDSCoordinate sw                        = corners[0];
-        NDSCoordinate se                        = corners[1];
-        NDSCoordinate ne                        = corners[2];
-        NDSCoordinate nw                        = corners[3];
+        // store the master tile corners
+        NDSBBox bb                              = t.getBBox();
+        NDSCoordinate sw                        = bb.southWest();
+        NDSCoordinate se                        = bb.southEast();
+        NDSCoordinate ne                        = bb.northEast();
+        NDSCoordinate nw                        = bb.northWest();
         // first, let's create four points that we know are in the child tiles
         int childTileLevel                      = masterTileLevel + 1;
-        NDSCoordinate csw                       = sw.getMidpoint(c);
-        NDSCoordinate cse                       = se.getMidpoint(c);
-        NDSCoordinate cne                       = ne.getMidpoint(c);
-        NDSCoordinate cnw                       = nw.getMidpoint(c);
+        NDSCoordinate csw                       = sw.fractionBetween(c, 0.5);
+        NDSCoordinate cse                       = se.fractionBetween(c, 0.5);
+        NDSCoordinate cne                       = ne.fractionBetween(c, 0.5);
+        NDSCoordinate cnw                       = nw.fractionBetween(c, 0.5);
         // get child tiles based on master level for double-checking
         NDSTile tsw                             = new NDSTile(childTileLevel, csw);
         NDSTile tse                             = new NDSTile(childTileLevel, cse);
         NDSTile tne                             = new NDSTile(childTileLevel, cne);
         NDSTile tnw                             = new NDSTile(childTileLevel, cnw);
-        int swNumber                            = t.getChildTileNumberSouthWest();
-        int seNumber                            = t.getChildTileNumberSouthEast();
-        int neNumber                            = t.getChildTileNumberNorthEast();
-        int nwNumber                            = t.getChildTileNumberNorthWest();
         // now perform check
-        assertEquals(tsw.getTileNumber(), swNumber);
-        assertEquals(tse.getTileNumber(), seNumber);
-        assertEquals(tne.getTileNumber(), neNumber);
-        assertEquals(tnw.getTileNumber(), nwNumber);
+        assertEquals(tsw.getTileNumber(), t.getChild(0).getTileNumber());
+        assertEquals(tse.getTileNumber(), t.getChild(1).getTileNumber());
+        assertEquals(tne.getTileNumber(), t.getChild(2).getTileNumber());
+        assertEquals(tnw.getTileNumber(), t.getChild(3).getTileNumber());
     }
 
     @Test
@@ -458,7 +367,7 @@ public class NDSTileTest {
         // set some random tile number and level for testing
         int tstTileNumber                       = 7;
         int tstTileLevel                        = 3;
-        int[] tstTileXY                         = masterTile.getTileXYfromTileID(tstTileLevel, tstTileNumber);
+        int[] tstTileXY                         = masterTile.getTileXYfromTileNumber(tstTileLevel, tstTileNumber);
         int tstTileX                            = tstTileXY[0];
         int tstTileY                            = tstTileXY[1];
         assertEquals(11, tstTileX);
@@ -471,49 +380,7 @@ public class NDSTileTest {
         tstTileXY                               = masterTile.getTileXYfromTileQuadkey(tstTileQuadkey);
         assertEquals(tstTileX, tstTileXY[0]);
         assertEquals(tstTileY, tstTileXY[1]);
-        assertEquals(tstTileNumber, masterTile.getTileIDfromTileXY(tstTileLevel, tstTileX, tstTileY));
-    }
-    
-    @Test
-    public void testEnvelopeWorks() {
-        // rough polygon data approximating Germany
-        double[][] polygonCoordinates = new double[][] {
-                {10.5,  45.9},
-                {13.0,  45.9},
-                {14.0,  49.0},
-                {12.0,  50.0},
-                {15.0,  51.0},
-                {15.0,  54.0},
-                {13.5,  54.5},
-                {11.0,  54.0},
-                {10.0,  55.0},
-                { 8.5,  55.0},
-                { 9.0,  54.0},
-                { 7.0,  53.5},
-                { 6.0,  52.0},
-                { 6.1,  50.0},
-                { 8.0,  49.0},
-                { 7.5,  47.5},
-                {10.5,  45.9}
-        };
-        NDSEnvelope envelope                = new NDSEnvelope(polygonCoordinates);
-        OctagonalEnvelope vividEnvelope     = envelope.getEnvelope();
-        // check envelope boundaries
-        assertEquals( 6.0, vividEnvelope.getMinX(), eps);
-        assertEquals(15.0, vividEnvelope.getMaxX(), eps);
-        assertEquals(45.9, vividEnvelope.getMinY(), eps);
-        assertEquals(55.0, vividEnvelope.getMaxY(), eps);
-        // check bounding box corners
-        assertEquals(581131592357515410L, envelope.getSouthWest().getMortonCode());
-        assertEquals(595825689965249734L, envelope.getSouthEast().getMortonCode());
-        assertEquals(592806849050488888L, envelope.getNorthEast().getMortonCode());
-        assertEquals(607500946658223212L, envelope.getNorthWest().getMortonCode());
-        // check master tile details
-        int[] info                          = envelope.getMasterTileInfo(15);
-        int level                           = info[0];
-        int number                          = info[1];
-        assertEquals(3, level);
-        assertEquals(8, number);
+        assertEquals(tstTileNumber, masterTile.getTileNumberFromTileXY(tstTileLevel, tstTileX, tstTileY));
     }
 
     @Test
